moved `toReceipt` into transport layer logic, moved `toPoints` into separate mapper file
